+++ ydb/library/benchmarks/queries/tpcds/pg/q32.sql
@@ -14,7 +14,7 @@ and d_date_sk = cs_sold_date_sk
 and cs_ext_discount_amt
      > (
          select
-            1.3::numeric * avg(cs_ext_discount_amt)
+            1.3 * avg(cs_ext_discount_amt)
          from
             {{catalog_sales}}
            ,{{date_dim}}
